refactor: 縮小 queryClient 作用域並優化樂觀更新邏輯
- 將 queryClient 限制在實際使用的 optimistic update/rollback 範圍內
- 簡化樂觀更新的樣板代碼，突出核心業務邏輯
- 重新命名 repoList 變數以符合其實際職責
